damage range from 10 - 20 dmg

+++ Assets/Scripts/Enemy_Scripts/EnemyMelee_AI_Health.cs
@@ -5,7 +5,7 @@ public class EnemyMelee_AI_Health : MonoBehaviour {
 
 	public int startingHealth = 100;
 	public int alliedArmour = 10;
-	public int currentHealth;
+	public float currentHealth;
 	public int moneyValue = 5;
     GameObject unitManager;
     UnitArrays Uarray;
@@ -48,7 +48,7 @@ public class EnemyMelee_AI_Health : MonoBehaviour {
 
 	}
 
-	public void TakeDamage (int damageTaken)
+	public void TakeDamage (float damageTaken)
 	{
 		if (Died == true) 
 		{
